LayoutTests/inspector/debugger/break-on-uncaught-exception-throw-in-promise-then-expected.txt

CONSOLE MESSAGE: line 55: testThrowingInPromiseThen
CONSOLE MESSAGE: line 57: in promise
CONSOLE MESSAGE: line 60: in promise.then
CONSOLE MESSAGE: line 7: throwing TestError
Checking pause locations when pausing on uncaught exceptions.

PAUSE #1 AT: doThrow:8:22
